move hover and measurement_basis callbacks
